questao-3-busca/requirements.txt
faiss-cpu==1.9.0
sentence-transformers==3.4.1
numpy==2.2.3
datasets==3.2.0
pytest
